Move heads and tails tests to their own file
